+++ tiger-maven-plugin/src/test/java/de/gematik/test/tiger/maven/adapter/mojos/TigerSerenityReportMojoTest.java
@@ -12,7 +12,6 @@ import static org.hamcrest.Matchers.startsWith;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.verify;
-
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -21,6 +20,7 @@ import org.apache.commons.io.FileUtils;
 import org.apache.maven.plugin.logging.Log;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
@@ -90,10 +90,25 @@ class TigerSerenityReportMojoTest {
     );
   }
 
+  @Tag("de.gematik.test.tiger.common.LongrunnerTest")
+  @Test
+  @DisplayName("If the property openSerenityReportInBrowser is set, the browser should open with the serenity report")
+  @SneakyThrows
+  void testIfTheBrowserOpensWithSerenityReport() {
+    // Preparation
+    underTest.setOpenSerenityReportInBrowser(true);
+    // Execution
+    underTest.execute();
+
+    // Assertion
+    verify(log).info("Browser for serenity report started");
+  }
+
   @SneakyThrows
   private void prepareReportDir() {
     final var repoRessourceDir = Paths.get(
         getClass().getResource("/serenetyReports/fresh").toURI());
     FileUtils.copyDirectory(repoRessourceDir.toFile(), reportDir.toFile());
   }
+
 }